chore: renames and restructure
