+++ app/src/features/apiClient/screens/apiClient/components/clientView/components/request/components/AuthorizationView/defaults.ts
@@ -1,0 +1,13 @@
+import { RQAPI } from "features/apiClient/types";
+import { Authorization } from "./types/AuthConfig";
+
+export function getDefaultAuthType(isRootRecord: boolean): Authorization.Type {
+  return isRootRecord ? Authorization.Type.NO_AUTH : Authorization.Type.INHERIT;
+}
+
+export function getDefaultAuth(isRootRecord: boolean): RQAPI.Auth {
+  return {
+    currentAuthType: getDefaultAuthType(isRootRecord),
+    authConfigStore: {},
+  };
+}
